ISSUE-157: Split up activity into read and write model
