+++ src/Domain/Strava/Activity/DistanceBreakdown.php
@@ -6,6 +6,8 @@ namespace App\Domain\Strava\Activity;
 
 use App\Domain\Measurement\Length\Kilometer;
 use App\Domain\Measurement\UnitSystem;
+use App\Domain\Strava\Activity\WriteModel\Activities;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 use Carbon\CarbonInterval;
 
 final readonly class DistanceBreakdown
