+++ src/Domain/Strava/Activity/Image/Image.php
@@ -2,7 +2,7 @@
 
 namespace App\Domain\Strava\Activity\Image;
 
-use App\Domain\Strava\Activity\Activity;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 
 final readonly class Image
 {
