+++ src/Domain/Strava/Activity/Stream/HeartRate.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Domain\Strava\Activity\Stream;
 
-use App\Domain\Strava\Activity\Activity;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 
 final readonly class HeartRate
 {
