+++ src/Domain/Strava/Activity/Stream/PowerOutput.php
@@ -2,7 +2,7 @@
 
 namespace App\Domain\Strava\Activity\Stream;
 
-use App\Domain\Strava\Activity\Activity;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 
 final readonly class PowerOutput
 {
