+++ src/Domain/Strava/Challenge/Consistency/ChallengeConsistency.php
@@ -6,9 +6,9 @@ namespace App\Domain\Strava\Challenge\Consistency;
 
 use App\Domain\Measurement\Length\Kilometer;
 use App\Domain\Measurement\Length\Meter;
-use App\Domain\Strava\Activity\Activities;
-use App\Domain\Strava\Activity\Activity;
 use App\Domain\Strava\Activity\ActivityType;
+use App\Domain\Strava\Activity\WriteModel\Activities;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 use App\Domain\Strava\Calendar\Months;
 use App\Domain\Strava\Calendar\Week;
