+++ src/Domain/Strava/Gear/GearStatistics.php
@@ -5,8 +5,8 @@ namespace App\Domain\Strava\Gear;
 use App\Domain\Measurement\Length\Kilometer;
 use App\Domain\Measurement\Length\Meter;
 use App\Domain\Measurement\Velocity\KmPerHour;
-use App\Domain\Strava\Activity\Activities;
-use App\Domain\Strava\Activity\Activity;
+use App\Domain\Strava\Activity\WriteModel\Activities;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 use Carbon\CarbonInterval;
 
 final readonly class GearStatistics
