+++ src/Domain/Strava/MonthlyStatistics.php
@@ -4,9 +4,9 @@ namespace App\Domain\Strava;
 
 use App\Domain\Measurement\Length\Kilometer;
 use App\Domain\Measurement\Length\Meter;
-use App\Domain\Strava\Activity\Activities;
-use App\Domain\Strava\Activity\Activity;
 use App\Domain\Strava\Activity\SportType\SportType;
+use App\Domain\Strava\Activity\WriteModel\Activities;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 use App\Domain\Strava\Calendar\Month;
 use App\Domain\Strava\Calendar\Months;
 use App\Domain\Strava\Challenge\Challenge;
