+++ tests/Domain/Strava/Activity/ActivityBuilder.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Tests\Domain\Strava\Activity;
 
-use App\Domain\Strava\Activity\Activity;
 use App\Domain\Strava\Activity\ActivityId;
 use App\Domain\Strava\Activity\SportType\SportType;
+use App\Domain\Strava\Activity\WriteModel\Activity;
 use App\Domain\Strava\Gear\GearId;
 use App\Infrastructure\Geocoding\Nominatim\Location;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
