+++ tests/Domain/Strava/Activity/ActivityTotalsTest.php
@@ -2,8 +2,8 @@
 
 namespace App\Tests\Domain\Strava\Activity;
 
-use App\Domain\Strava\Activity\Activities;
 use App\Domain\Strava\Activity\ActivityTotals;
+use App\Domain\Strava\Activity\WriteModel\Activities;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
