+++ tests/Domain/Strava/Gear/ImportGear/ImportGearCommandHandlerTest.php
@@ -3,7 +3,7 @@
 namespace App\Tests\Domain\Strava\Gear\ImportGear;
 
 use App\Domain\Strava\Activity\ActivityId;
-use App\Domain\Strava\Activity\ActivityRepository;
+use App\Domain\Strava\Activity\WriteModel\ActivityRepository;
 use App\Domain\Strava\Gear\GearId;
 use App\Domain\Strava\Gear\GearRepository;
 use App\Domain\Strava\Gear\ImportGear\ImportGear;
